refactor: replace checkOclConstraints with validateDiagram for unified validation approach and update related components
